requirements.txt
Flask==1.1.2
pyqrcode==1.2.1
pypng==0.0.20